refactor: Complete authentication refactoring implementation
- Removed backward compatibility code from initializeProviders.ts
- Deleted obsolete migrationHelper.ts
- Consolidated AuthProviderRegistry into AuthProviderFactory for simpler API
- Removed legacy config fields (webClientID, tenantID, auth0Domain, etc.)
- Updated all imports to use the consolidated AuthProviderFactory
- Replaced refactored files with actual implementation
- All tests passing, build successful

This completes the simplification of the N-provider authentication architecture
by removing ~150 lines of backward compatibility code and consolidating the
registry/factory pattern into a single, cleaner interface.
